Move logic needed by other components out of PackageService into CorePackageService (#4721)
Add the UpdatePackageStatusAsync method for package status transitions
