+++ src/NuGetGallery.Core/Services/CorePackageService.cs
@@ -1,0 +1,215 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+using System.Collections.Generic;
+using System.Data.Entity;
+using System.Linq;
+using System.Threading.Tasks;
+using NuGet.Versioning;
+
+namespace NuGetGallery
+{
+    public class CorePackageService : ICorePackageService
+    {
+        protected readonly IEntityRepository<Package> _packageRepository;
+        
+        public CorePackageService(IEntityRepository<Package> packageRepository)
+        {
+            _packageRepository = packageRepository;
+        }
+
+        public virtual async Task UpdatePackageStatusAsync(
+            Package package,
+            PackageStatus newPackageStatus,
+            bool commitChanges = true)
+        {
+            if (package == null)
+            {
+                throw new ArgumentNullException(nameof(package));
+            }
+
+            if (package.PackageRegistration == null)
+            {
+                throw new ArgumentException($"The {nameof(Package.PackageRegistration)} property must populated on the provided package.", nameof(package));
+            }
+
+            // Avoid all of this work if the package status is not changing.
+            if (package.PackageStatusKey != newPackageStatus)
+            {
+#pragma warning disable CS0612 // Type or member is obsolete
+                package.Deleted = newPackageStatus == PackageStatus.Deleted;
+#pragma warning restore CS0612 // Type or member is obsolete
+                package.PackageStatusKey = newPackageStatus;
+                package.LastUpdated = DateTime.UtcNow;
+
+                /// If the package is being made available, consider this change as an "edit" so that the package
+                /// appears for cursors on the <see cref="Package.LastEdited"/> field.
+                if (newPackageStatus == PackageStatus.Available)
+                {
+                    package.LastEdited = DateTime.UtcNow;
+                }
+
+                // If the package is just now becoming available or if it was previously a latest package, then
+                // re-evaluate all of the latest bits.
+                if (newPackageStatus == PackageStatus.Available ||
+                    package.IsLatest ||
+                    package.IsLatestStable ||
+                    package.IsLatestSemVer2 ||
+                    package.IsLatestStableSemVer2)
+                {
+                    await UpdateIsLatestAsync(package.PackageRegistration, commitChanges: false);
+                }
+            }
+
+            if (commitChanges)
+            {
+                // Even if the package status did not change, commit the changes in the entity context to pick up any
+                // other entity changes.
+                await _packageRepository.CommitChangesAsync();
+            }
+        }
+
+        public virtual async Task UpdateIsLatestAsync(PackageRegistration packageRegistration, bool commitChanges = true)
+        {
+            if (!packageRegistration.Packages.Any())
+            {
+                return;
+            }
+
+            // TODO: improve setting the latest bit; this is horrible. Trigger maybe?
+            var currentUtcTime = DateTime.UtcNow;
+            foreach (var pv in packageRegistration.Packages.Where(p => p.IsLatest || p.IsLatestStable || p.IsLatestSemVer2 || p.IsLatestStableSemVer2))
+            {
+                pv.IsLatest = false;
+                pv.IsLatestStable = false;
+                pv.IsLatestSemVer2 = false;
+                pv.IsLatestStableSemVer2 = false;
+                pv.LastUpdated = currentUtcTime;
+            }
+
+            // If the last listed package was just unlisted, then we won't find another one
+            var latestPackage = FindPackage(
+                packageRegistration.Packages.AsQueryable(),
+                ps => ps
+                    .Where(SemVerLevelKey.IsUnknownPredicate())
+                    .Where(p => p.PackageStatusKey == PackageStatus.Available && p.Listed));
+
+            var latestSemVer2Package = FindPackage(
+                packageRegistration.Packages.AsQueryable(),
+                ps => ps
+                    .Where(SemVerLevelKey.IsSemVer2Predicate())
+                    .Where(p => p.PackageStatusKey == PackageStatus.Available && p.Listed));
+
+            if (latestPackage != null)
+            {
+                latestPackage.IsLatest = true;
+                latestPackage.LastUpdated = currentUtcTime;
+
+                if (latestPackage.IsPrerelease)
+                {
+                    // If the newest uploaded package is a prerelease package, we need to find an older package that is
+                    // a release version and set it to IsLatest.
+                    var latestReleasePackage = FindPackage(packageRegistration
+                        .Packages
+                        .AsQueryable()
+                        .Where(SemVerLevelKey.IsUnknownPredicate())
+                        .Where(p => !p.IsPrerelease && p.PackageStatusKey == PackageStatus.Available && p.Listed));
+
+                    if (latestReleasePackage != null)
+                    {
+                        // We could have no release packages
+                        latestReleasePackage.IsLatestStable = true;
+                        latestReleasePackage.LastUpdated = currentUtcTime;
+                    }
+                }
+                else
+                {
+                    // Only release versions are marked as IsLatestStable.
+                    latestPackage.IsLatestStable = true;
+                }
+            }
+
+            if (latestSemVer2Package != null)
+            {
+                latestSemVer2Package.IsLatestSemVer2 = true;
+                latestSemVer2Package.LastUpdated = currentUtcTime;
+
+                if (latestSemVer2Package.IsPrerelease)
+                {
+                    // If the newest uploaded package is a prerelease package, we need to find an older package that is
+                    // a release version and set it to IsLatest.
+                    var latestSemVer2ReleasePackage = FindPackage(packageRegistration
+                        .Packages
+                        .AsQueryable()
+                        .Where(SemVerLevelKey.IsSemVer2Predicate())
+                        .Where(p => !p.IsPrerelease && p.PackageStatusKey == PackageStatus.Available && p.Listed));
+
+                    if (latestSemVer2ReleasePackage != null)
+                    {
+                        // We could have no release packages
+                        latestSemVer2ReleasePackage.IsLatestStableSemVer2 = true;
+                        latestSemVer2ReleasePackage.LastUpdated = currentUtcTime;
+                    }
+                }
+                else
+                {
+                    // Only release versions are marked as IsLatestStable.
+                    latestSemVer2Package.IsLatestStableSemVer2 = true;
+                }
+            }
+
+            if (commitChanges)
+            {
+                await _packageRepository.CommitChangesAsync();
+            }
+        }
+
+        public virtual Package FindPackageByIdAndVersionStrict(string id, string version)
+        {
+            if (string.IsNullOrWhiteSpace(id))
+            {
+                throw new ArgumentNullException(nameof(id));
+            }
+
+            if (string.IsNullOrEmpty(version))
+            {
+                throw new ArgumentException(nameof(version));
+            }
+
+            var normalizedVersion = NuGetVersionFormatter.Normalize(version);
+
+            // These string comparisons are case-(in)sensitive depending on SQLServer collation.
+            // Case-insensitive collation is recommended, e.g. SQL_Latin1_General_CP1_CI_AS.
+            var package = GetPackagesByIdQueryable(id)
+                .SingleOrDefault(p => p.NormalizedVersion == normalizedVersion);
+
+            return package;
+        }
+
+        protected IQueryable<Package> GetPackagesByIdQueryable(string id)
+        {
+            return _packageRepository
+                .GetAll()
+                .Include(p => p.LicenseReports)
+                .Include(p => p.PackageRegistration)
+                .Where(p => p.PackageRegistration.Id == id);
+        }
+
+        private static Package FindPackage(IQueryable<Package> packages, Func<IQueryable<Package>, IQueryable<Package>> predicate = null)
+        {
+            if (predicate != null)
+            {
+                packages = predicate(packages);
+            }
+
+            NuGetVersion version = packages.Max(p => new NuGetVersion(p.Version));
+            if (version == null)
+            {
+                return null;
+            }
+
+            return packages.First(pv => pv.Version.Equals(version.OriginalVersion, StringComparison.OrdinalIgnoreCase));
+        }
+    }
+}
