+++ tests/NuGetGallery.Core.Facts/Services/CorePackageServiceFacts.cs
@@ -1,0 +1,454 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using Moq;
+using Xunit;
+
+namespace NuGetGallery
+{
+    public class CorePackageServiceFacts
+    {
+        public class TheUpdatePackageStatusMethod
+        {
+            [Fact]
+            public async Task RejectsNullPackage()
+            {
+                // Arrange
+                Package package = null;
+                var service = CreateService();
+
+                // Act & Assert
+                await Assert.ThrowsAsync<ArgumentNullException>(
+                    () => service.UpdatePackageStatusAsync(package, PackageStatus.Deleted, commitChanges: true));
+            }
+
+            [Fact]
+            public async Task RejectsNullPackageRegistration()
+            {
+                // Arrange
+                Package package = new Package
+                {
+                    PackageRegistration = null,
+                };
+                var service = CreateService();
+
+                // Act & Assert
+                await Assert.ThrowsAsync<ArgumentException>(
+                    () => service.UpdatePackageStatusAsync(package, PackageStatus.Deleted, commitChanges: true));
+            }
+
+            [Theory]
+            [InlineData(false, 0)]
+            [InlineData(true, 1)]
+            public async Task CommitsTheCorrectNumberOfTimes(bool commitChanges, int commitCount)
+            {
+                // Arrange
+                var packageRepositoy = new Mock<IEntityRepository<Package>>();
+                var package = new Package
+                {
+                    PackageRegistration = new PackageRegistration(),
+                };
+                var service = CreateService(packageRepository: packageRepositoy);
+
+                // Act
+                await service.UpdatePackageStatusAsync(package, PackageStatus.Available, commitChanges);
+
+                // Assert
+                packageRepositoy.Verify(x => x.CommitChangesAsync(), Times.Exactly(commitCount));
+            }
+
+            [Fact]
+            public async Task SetDeletedBitWhenNewPackageStatusIsAvailable()
+            {
+                // Arrange
+                var service = CreateService();
+                var package = new Package
+                {
+                    PackageStatusKey = PackageStatus.Available,
+#pragma warning disable CS0612 // Type or member is obsolete
+                    Deleted = false,
+#pragma warning restore CS0612 // Type or member is obsolete
+                    PackageRegistration = new PackageRegistration
+                    {
+                        Packages = new List<Package>(),
+                    }
+                };
+                package.PackageRegistration.Packages.Add(package);
+
+                // Act
+                await service.UpdatePackageStatusAsync(package, PackageStatus.Deleted);
+
+                // Assert
+#pragma warning disable CS0612 // Type or member is obsolete
+                Assert.True(package.Deleted);
+#pragma warning restore CS0612 // Type or member is obsolete
+            }
+
+            [Fact]
+            public async Task UpdateCorrectPropertiesWhenBecomingAvailable()
+            {
+                // Arrange
+                var service = CreateService();
+                var package = new Package
+                {
+                    Version = "1.0.0",
+                    PackageStatusKey = PackageStatus.Deleted,
+#pragma warning disable CS0612 // Type or member is obsolete
+                    Deleted = true,
+#pragma warning restore CS0612 // Type or member is obsolete
+                    PackageRegistration = new PackageRegistration
+                    {
+                        Packages = new List<Package>(),
+                    },
+                    IsLatest = false,
+                    IsLatestStable = false,
+                    IsLatestSemVer2 = false,
+                    IsLatestStableSemVer2 = false,
+                    SemVerLevelKey = SemVerLevelKey.Unknown,
+                    LastEdited = DateTime.MinValue,
+                    LastUpdated = DateTime.MinValue,
+                };
+                package.PackageRegistration.Packages.Add(package);
+
+                // Act
+                var before = DateTime.UtcNow;
+                await service.UpdatePackageStatusAsync(package, PackageStatus.Available);
+                var after = DateTime.UtcNow;
+
+                // Assert
+#pragma warning disable CS0612 // Type or member is obsolete
+                Assert.False(package.Deleted, $"{nameof(Package.Deleted)} should be false.");
+#pragma warning restore CS0612 // Type or member is obsolete
+                Assert.Equal(PackageStatus.Available, package.PackageStatusKey);
+                Assert.True(package.IsLatest, $"{nameof(Package.IsLatest)} should be true.");
+                Assert.True(package.IsLatestStable, $"{nameof(Package.IsLatestStable)} should be true.");
+                Assert.True(package.IsLatestSemVer2, $"{nameof(Package.IsLatestSemVer2)} should be true.");
+                Assert.True(package.IsLatestStableSemVer2, $"{nameof(Package.IsLatestStableSemVer2)} should be true.");
+                Assert.NotNull(package.LastEdited);
+                Assert.InRange(package.LastEdited.Value, before, after);
+                Assert.InRange(package.LastUpdated, before, after);
+            }
+
+            [Fact]
+            public async Task DoesNotUpdateLatestPropertiesWhenNoLatestAndNotBecomingAvailable()
+            {
+                // Arrange
+                var service = CreateService();
+                var package = new Package
+                {
+                    PackageStatusKey = PackageStatus.Available,
+#pragma warning disable CS0612 // Type or member is obsolete
+                    Deleted = false,
+#pragma warning restore CS0612 // Type or member is obsolete
+                    PackageRegistration = new PackageRegistration
+                    {
+                        Packages = new List<Package>(),
+                    },
+                    IsLatest = false,
+                    IsLatestStable = false,
+                    IsLatestSemVer2 = false,
+                    IsLatestStableSemVer2 = false,
+                    SemVerLevelKey = SemVerLevelKey.Unknown,
+                };
+                package.PackageRegistration.Packages.Add(package);
+
+                // Act
+                var before = DateTime.UtcNow;
+                await service.UpdatePackageStatusAsync(package, PackageStatus.Deleted);
+                var after = DateTime.UtcNow;
+
+                // Assert
+                Assert.False(package.IsLatest, $"{nameof(Package.IsLatest)} should be false.");
+                Assert.False(package.IsLatestStable, $"{nameof(Package.IsLatestStable)} should be false.");
+                Assert.False(package.IsLatestSemVer2, $"{nameof(Package.IsLatestSemVer2)} should be false.");
+                Assert.False(package.IsLatestStableSemVer2, $"{nameof(Package.IsLatestStableSemVer2)} should be false.");
+                Assert.Null(package.LastEdited);
+                Assert.InRange(package.LastUpdated, before, after);
+            }
+
+            [Fact]
+            public async Task DoesNotUpdateLastEditedWhenBecomingUnavailable()
+            {
+                // Arrange
+                var service = CreateService();
+                var package = new Package
+                {
+                    PackageStatusKey = PackageStatus.Available,
+                    PackageRegistration = new PackageRegistration
+                    {
+                        Packages = new List<Package>(),
+                    },
+                };
+                package.PackageRegistration.Packages.Add(package);
+
+                // Act
+                await service.UpdatePackageStatusAsync(package, PackageStatus.Deleted);
+
+                // Assert
+                Assert.Null(package.LastEdited);
+            }
+        }
+
+        public class TheUpdateIsLatestMethod
+        {
+            [Fact]
+            public async Task DoNotCommitIfCommitChangesIsFalse()
+            {
+                // Arrange
+                var packageRegistration = new PackageRegistration();
+                var package = new Package { PackageRegistration = packageRegistration, Version = "1.0.0" };
+                packageRegistration.Packages.Add(package);
+                var packageRepository = new Mock<IEntityRepository<Package>>();
+
+                var service = CreateService(packageRepository: packageRepository);
+
+                // Act
+                await service.UpdateIsLatestAsync(packageRegistration, commitChanges: false);
+
+                // Assert
+                packageRepository.Verify(r => r.CommitChangesAsync(), Times.Never());
+            }
+
+            [Fact]
+            public async Task CommitIfCommitChangesIsTrue()
+            {
+                // Arrange
+                var packageRegistration = new PackageRegistration();
+                var package = new Package { PackageRegistration = packageRegistration, Version = "1.0.0" };
+                packageRegistration.Packages.Add(package);
+                var packageRepository = new Mock<IEntityRepository<Package>>();
+
+                var service = CreateService(packageRepository: packageRepository);
+
+                // Act
+                await service.UpdateIsLatestAsync(packageRegistration, true);
+
+                // Assert
+                packageRepository.Verify(r => r.CommitChangesAsync(), Times.Once());
+            }
+
+            [Fact]
+            public async Task ResetsCurrentLatestPackageVersionsBeforeUpdate()
+            {
+                // Arrange
+                var packageRegistration = new PackageRegistration();
+
+                var previousLatestStable = new Package { PackageRegistration = packageRegistration, Version = "1.0.0", IsLatestStable = true };
+                packageRegistration.Packages.Add(previousLatestStable);
+                var previousLatest = new Package { PackageRegistration = packageRegistration, Version = "1.0.0-alpha", IsLatest = true, IsPrerelease = true };
+                packageRegistration.Packages.Add(previousLatest);
+                var previousLatestStableSemVer2 = new Package { PackageRegistration = packageRegistration, Version = "1.0.1+metadata", IsLatestStableSemVer2 = true, SemVerLevelKey = SemVerLevelKey.SemVer2 };
+                packageRegistration.Packages.Add(previousLatestStableSemVer2);
+                var previousLatestSemVer2 = new Package { PackageRegistration = packageRegistration, Version = "1.0.1-alpha.1", IsLatestSemVer2 = true, IsPrerelease = true, SemVerLevelKey = SemVerLevelKey.SemVer2 };
+                packageRegistration.Packages.Add(previousLatestSemVer2);
+
+                // Simulates adding newer versions, to ensure the previous latest are no longer latest at end of test.
+                var newLatestStable = new Package { PackageRegistration = packageRegistration, Version = "1.0.1", IsLatestStable = true };
+                packageRegistration.Packages.Add(newLatestStable);
+                var newLatest = new Package { PackageRegistration = packageRegistration, Version = "1.0.2-alpha", IsLatest = true, IsPrerelease = true };
+                packageRegistration.Packages.Add(newLatest);
+                var newLatestStableSemVer2 = new Package { PackageRegistration = packageRegistration, Version = "1.0.2+metadata", IsLatestStableSemVer2 = true, SemVerLevelKey = SemVerLevelKey.SemVer2 };
+                packageRegistration.Packages.Add(newLatestStableSemVer2);
+                var newLatestSemVer2 = new Package { PackageRegistration = packageRegistration, Version = "1.0.3-alpha.1", IsLatestSemVer2 = true, IsPrerelease = true, SemVerLevelKey = SemVerLevelKey.SemVer2 };
+                packageRegistration.Packages.Add(newLatestSemVer2);
+
+                var packageRepository = new Mock<IEntityRepository<Package>>();
+
+                var service = CreateService(packageRepository: packageRepository);
+
+                // Act
+                await service.UpdateIsLatestAsync(packageRegistration, commitChanges: true);
+
+                // Assert
+                Assert.False(previousLatestStable.IsLatestStable);
+                Assert.False(previousLatest.IsLatest);
+                Assert.False(previousLatestSemVer2.IsLatestSemVer2);
+                Assert.False(previousLatestStableSemVer2.IsLatestStableSemVer2);
+
+                Assert.True(newLatestStable.IsLatestStable);
+                Assert.True(newLatest.IsLatest);
+                Assert.True(newLatestSemVer2.IsLatestSemVer2);
+                Assert.True(newLatestStableSemVer2.IsLatestStableSemVer2);
+            }
+
+            [Fact]
+            public async Task UpdateIsLatestScenarioForPrereleaseAsAbsoluteLatest()
+            {
+                // Arrange
+                var packages = new HashSet<Package>();
+                var packageRegistration = new PackageRegistration { Packages = packages };
+                var package10A = new Package { PackageRegistration = packageRegistration, Version = "1.0.0-a", IsPrerelease = true };
+                packages.Add(package10A);
+                var package09 = new Package { PackageRegistration = packageRegistration, Version = "0.9.0" };
+                packages.Add(package09);
+                var packageRepository = new Mock<IEntityRepository<Package>>(MockBehavior.Strict);
+                packageRepository.Setup(r => r.CommitChangesAsync())
+                    .Returns(Task.FromResult(0)).Verifiable();
+                var service = CreateService(packageRepository: packageRepository);
+
+                // Act
+                await service.UpdateIsLatestAsync(packageRegistration, true);
+
+                // Assert
+                Assert.True(package10A.IsLatest);
+                Assert.True(package10A.IsLatestSemVer2);
+                Assert.False(package10A.IsLatestStable);
+                Assert.False(package10A.IsLatestStableSemVer2);
+
+                Assert.False(package09.IsLatest);
+                Assert.False(package09.IsLatestSemVer2);
+                Assert.True(package09.IsLatestStable);
+                Assert.True(package09.IsLatestStableSemVer2);
+
+                packageRepository.Verify();
+            }
+
+            [Fact]
+            public async Task UpdateIsLatestScenarioForStableAsAbsoluteLatest()
+            {
+                // Arrange
+                var packages = new HashSet<Package>();
+                var packageRegistration = new PackageRegistration { Packages = packages };
+                var package100 = new Package { PackageRegistration = packageRegistration, Version = "1.0.0" };
+                packages.Add(package100);
+                var package10A = new Package { PackageRegistration = packageRegistration, Version = "1.0.0-a", IsPrerelease = true };
+                packages.Add(package10A);
+                var package09 = new Package { PackageRegistration = packageRegistration, Version = "0.9.0" };
+                packages.Add(package09);
+                var packageRepository = new Mock<IEntityRepository<Package>>(MockBehavior.Strict);
+                packageRepository.Setup(r => r.CommitChangesAsync())
+                    .Returns(Task.FromResult(0)).Verifiable();
+                var service = CreateService(packageRepository: packageRepository);
+
+                // Act
+                await service.UpdateIsLatestAsync(packageRegistration, true);
+
+                // Assert
+                Assert.True(package100.IsLatest);
+                Assert.True(package100.IsLatestSemVer2);
+                Assert.True(package100.IsLatestStable);
+                Assert.True(package100.IsLatestStableSemVer2);
+
+                Assert.False(package10A.IsLatest);
+                Assert.False(package10A.IsLatestSemVer2);
+                Assert.False(package10A.IsLatestStable);
+                Assert.False(package10A.IsLatestStableSemVer2);
+
+                Assert.False(package09.IsLatest);
+                Assert.False(package09.IsLatestSemVer2);
+                Assert.False(package09.IsLatestStable);
+                Assert.False(package09.IsLatestStableSemVer2);
+
+                packageRepository.Verify();
+            }
+
+            [Fact]
+            public async Task UpdateIsLatestScenarioForSemVer2PrereleaseAsAbsoluteLatest()
+            {
+                // Arrange
+                var packages = new HashSet<Package>();
+                var packageRegistration = new PackageRegistration { Packages = packages };
+                var semVer2Package = new Package { PackageRegistration = packageRegistration, Version = "1.0.1-alpha.1", IsPrerelease = true, SemVerLevelKey = SemVerLevelKey.SemVer2 };
+                packages.Add(semVer2Package);
+                var package100 = new Package { PackageRegistration = packageRegistration, Version = "1.0.0" };
+                packages.Add(package100);
+                var package10A = new Package { PackageRegistration = packageRegistration, Version = "1.0.0-a", IsPrerelease = true };
+                packages.Add(package10A);
+                var package09 = new Package { PackageRegistration = packageRegistration, Version = "0.9.0" };
+                packages.Add(package09);
+                var packageRepository = new Mock<IEntityRepository<Package>>(MockBehavior.Strict);
+                packageRepository.Setup(r => r.CommitChangesAsync())
+                    .Returns(Task.FromResult(0)).Verifiable();
+                var service = CreateService(packageRepository: packageRepository);
+
+                // Act
+                await service.UpdateIsLatestAsync(packageRegistration, true);
+
+                // Assert
+                Assert.True(semVer2Package.IsLatestSemVer2);
+                Assert.False(semVer2Package.IsLatestStableSemVer2);
+                Assert.False(semVer2Package.IsLatest);
+                Assert.False(semVer2Package.IsLatestStable);
+
+                Assert.True(package100.IsLatest);
+                Assert.False(package100.IsLatestSemVer2);
+                Assert.True(package100.IsLatestStable);
+                Assert.True(package100.IsLatestStableSemVer2);
+
+                Assert.False(package10A.IsLatest);
+                Assert.False(package10A.IsLatestSemVer2);
+                Assert.False(package10A.IsLatestStable);
+                Assert.False(package10A.IsLatestStableSemVer2);
+
+                Assert.False(package09.IsLatest);
+                Assert.False(package09.IsLatestSemVer2);
+                Assert.False(package09.IsLatestStable);
+                Assert.False(package09.IsLatestStableSemVer2);
+
+                packageRepository.Verify();
+            }
+
+            [Fact]
+            public async Task UpdateIsLatestScenarioForSemVer2StableAsAbsoluteLatest()
+            {
+                // Arrange
+                var packages = new HashSet<Package>();
+                var packageRegistration = new PackageRegistration { Packages = packages };
+                var semVer2Package = new Package { PackageRegistration = packageRegistration, Version = "1.0.1+metadata", SemVerLevelKey = SemVerLevelKey.SemVer2 };
+                packages.Add(semVer2Package);
+                var package100 = new Package { PackageRegistration = packageRegistration, Version = "1.0.0" };
+                packages.Add(package100);
+                var package10A = new Package { PackageRegistration = packageRegistration, Version = "1.0.0-a", IsPrerelease = true };
+                packages.Add(package10A);
+                var package09 = new Package { PackageRegistration = packageRegistration, Version = "0.9.0" };
+                packages.Add(package09);
+                var packageRepository = new Mock<IEntityRepository<Package>>(MockBehavior.Strict);
+                packageRepository.Setup(r => r.CommitChangesAsync())
+                    .Returns(Task.FromResult(0)).Verifiable();
+                var service = CreateService(packageRepository: packageRepository);
+
+                // Act
+                await service.UpdateIsLatestAsync(packageRegistration, true);
+
+                // Assert
+                Assert.True(semVer2Package.IsLatestSemVer2);
+                Assert.True(semVer2Package.IsLatestStableSemVer2);
+                Assert.False(semVer2Package.IsLatest);
+                Assert.False(semVer2Package.IsLatestStable);
+
+                Assert.True(package100.IsLatest);
+                Assert.False(package100.IsLatestSemVer2);
+                Assert.True(package100.IsLatestStable);
+                Assert.False(package100.IsLatestStableSemVer2);
+
+                Assert.False(package10A.IsLatest);
+                Assert.False(package10A.IsLatestSemVer2);
+                Assert.False(package10A.IsLatestStable);
+                Assert.False(package10A.IsLatestStableSemVer2);
+
+                Assert.False(package09.IsLatest);
+                Assert.False(package09.IsLatestSemVer2);
+                Assert.False(package09.IsLatestStable);
+                Assert.False(package09.IsLatestStableSemVer2);
+
+                packageRepository.Verify();
+            }
+        }
+
+        private static ICorePackageService CreateService(
+            Mock<IEntityRepository<Package>> packageRepository = null)
+        {
+            packageRepository = packageRepository ?? new Mock<IEntityRepository<Package>>();
+            
+            var packageService = new Mock<CorePackageService>(
+                packageRepository.Object);
+
+            packageService.CallBase = true;
+
+            return packageService.Object;
+        }
+    }
+}
