+++ tests/NuGetGallery.Facts/Services/PackageServiceFacts.cs
@@ -12,14 +12,13 @@ using NuGet.Versioning;
 using NuGetGallery.Auditing;
 using NuGetGallery.Framework;
 using NuGetGallery.Packaging;
-using Xunit;
 using NuGetGallery.TestUtils;
+using Xunit;
 
 namespace NuGetGallery
 {
     public class PackageServiceFacts
     {
-
         private static IPackageService CreateService(
             Mock<IEntityRepository<PackageRegistration>> packageRegistrationRepository = null,
             Mock<IEntityRepository<Package>> packageRepository = null,
@@ -885,252 +884,6 @@ namespace NuGetGallery
                 var request = await service.CreatePackageOwnerRequestAsync(package, owner, newOwner);
 
                 Assert.Equal(99, request.RequestingOwnerKey);
-            }
-        }
-
-        public class TheUpdateIsLatestMethod
-        {
-            [Fact]
-            public async Task DoNotCommitIfCommitChangesIsFalse()
-            {
-                // Arrange
-                var packageRegistration = new PackageRegistration();
-                var package = new Package { PackageRegistration = packageRegistration, Version = "1.0.0" };
-                packageRegistration.Packages.Add(package);
-                var packageRepository = new Mock<IEntityRepository<Package>>();
-
-                var service = CreateService(packageRepository: packageRepository);
-
-                // Act
-                await service.UpdateIsLatestAsync(packageRegistration, commitChanges: false);
-
-                // Assert
-                packageRepository.Verify(r => r.CommitChangesAsync(), Times.Never());
-            }
-
-            [Fact]
-            public async Task CommitIfCommitChangesIsTrue()
-            {
-                // Arrange
-                var packageRegistration = new PackageRegistration();
-                var package = new Package { PackageRegistration = packageRegistration, Version = "1.0.0" };
-                packageRegistration.Packages.Add(package);
-                var packageRepository = new Mock<IEntityRepository<Package>>();
-
-                var service = CreateService(packageRepository: packageRepository);
-
-                // Act
-                await service.UpdateIsLatestAsync(packageRegistration, true);
-
-                // Assert
-                packageRepository.Verify(r => r.CommitChangesAsync(), Times.Once());
-            }
-
-            [Fact]
-            public async Task ResetsCurrentLatestPackageVersionsBeforeUpdate()
-            {
-                // Arrange
-                var packageRegistration = new PackageRegistration();
-
-                var previousLatestStable = new Package { PackageRegistration = packageRegistration, Version = "1.0.0", IsLatestStable = true };
-                packageRegistration.Packages.Add(previousLatestStable);
-                var previousLatest = new Package { PackageRegistration = packageRegistration, Version = "1.0.0-alpha", IsLatest = true, IsPrerelease = true };
-                packageRegistration.Packages.Add(previousLatest);
-                var previousLatestStableSemVer2 = new Package { PackageRegistration = packageRegistration, Version = "1.0.1+metadata", IsLatestStableSemVer2 = true, SemVerLevelKey = SemVerLevelKey.SemVer2 };
-                packageRegistration.Packages.Add(previousLatestStableSemVer2);
-                var previousLatestSemVer2 = new Package { PackageRegistration = packageRegistration, Version = "1.0.1-alpha.1", IsLatestSemVer2 = true, IsPrerelease = true, SemVerLevelKey = SemVerLevelKey.SemVer2 };
-                packageRegistration.Packages.Add(previousLatestSemVer2);
-
-                // Simulates adding newer versions, to ensure the previous latest are no longer latest at end of test.
-                var newLatestStable = new Package { PackageRegistration = packageRegistration, Version = "1.0.1", IsLatestStable = true };
-                packageRegistration.Packages.Add(newLatestStable);
-                var newLatest = new Package { PackageRegistration = packageRegistration, Version = "1.0.2-alpha", IsLatest = true, IsPrerelease = true };
-                packageRegistration.Packages.Add(newLatest);
-                var newLatestStableSemVer2 = new Package { PackageRegistration = packageRegistration, Version = "1.0.2+metadata", IsLatestStableSemVer2 = true, SemVerLevelKey = SemVerLevelKey.SemVer2 };
-                packageRegistration.Packages.Add(newLatestStableSemVer2);
-                var newLatestSemVer2 = new Package { PackageRegistration = packageRegistration, Version = "1.0.3-alpha.1", IsLatestSemVer2 = true, IsPrerelease = true, SemVerLevelKey = SemVerLevelKey.SemVer2 };
-                packageRegistration.Packages.Add(newLatestSemVer2);
-
-                var packageRepository = new Mock<IEntityRepository<Package>>();
-
-                var service = CreateService(packageRepository: packageRepository);
-
-                // Act
-                await service.UpdateIsLatestAsync(packageRegistration, commitChanges: true);
-
-                // Assert
-                Assert.False(previousLatestStable.IsLatestStable);
-                Assert.False(previousLatest.IsLatest);
-                Assert.False(previousLatestSemVer2.IsLatestSemVer2);
-                Assert.False(previousLatestStableSemVer2.IsLatestStableSemVer2);
-                
-                Assert.True(newLatestStable.IsLatestStable);
-                Assert.True(newLatest.IsLatest);
-                Assert.True(newLatestSemVer2.IsLatestSemVer2);
-                Assert.True(newLatestStableSemVer2.IsLatestStableSemVer2);
-            }
-
-            [Fact]
-            public async Task UpdateIsLatestScenarioForPrereleaseAsAbsoluteLatest()
-            {
-                // Arrange
-                var packages = new HashSet<Package>();
-                var packageRegistration = new PackageRegistration { Packages = packages };
-                var package10A = new Package { PackageRegistration = packageRegistration, Version = "1.0.0-a", IsPrerelease = true };
-                packages.Add(package10A);
-                var package09 = new Package { PackageRegistration = packageRegistration, Version = "0.9.0" };
-                packages.Add(package09);
-                var packageRepository = new Mock<IEntityRepository<Package>>(MockBehavior.Strict);
-                packageRepository.Setup(r => r.CommitChangesAsync())
-                    .Returns(Task.CompletedTask).Verifiable();
-                var service = CreateService(packageRepository: packageRepository);
-
-                // Act
-                await service.UpdateIsLatestAsync(packageRegistration, true);
-
-                // Assert
-                Assert.True(package10A.IsLatest);
-                Assert.True(package10A.IsLatestSemVer2);
-                Assert.False(package10A.IsLatestStable);
-                Assert.False(package10A.IsLatestStableSemVer2);
-
-                Assert.False(package09.IsLatest);
-                Assert.False(package09.IsLatestSemVer2);
-                Assert.True(package09.IsLatestStable);
-                Assert.True(package09.IsLatestStableSemVer2);
-
-                packageRepository.Verify();
-            }
-
-            [Fact]
-            public async Task UpdateIsLatestScenarioForStableAsAbsoluteLatest()
-            {
-                // Arrange
-                var packages = new HashSet<Package>();
-                var packageRegistration = new PackageRegistration { Packages = packages };
-                var package100 = new Package { PackageRegistration = packageRegistration, Version = "1.0.0" };
-                packages.Add(package100);
-                var package10A = new Package { PackageRegistration = packageRegistration, Version = "1.0.0-a", IsPrerelease = true };
-                packages.Add(package10A);
-                var package09 = new Package { PackageRegistration = packageRegistration, Version = "0.9.0" };
-                packages.Add(package09);
-                var packageRepository = new Mock<IEntityRepository<Package>>(MockBehavior.Strict);
-                packageRepository.Setup(r => r.CommitChangesAsync())
-                    .Returns(Task.CompletedTask).Verifiable();
-                var service = CreateService(packageRepository: packageRepository);
-
-                // Act
-                await service.UpdateIsLatestAsync(packageRegistration, true);
-
-                // Assert
-                Assert.True(package100.IsLatest);
-                Assert.True(package100.IsLatestSemVer2);
-                Assert.True(package100.IsLatestStable);
-                Assert.True(package100.IsLatestStableSemVer2);
-
-                Assert.False(package10A.IsLatest);
-                Assert.False(package10A.IsLatestSemVer2);
-                Assert.False(package10A.IsLatestStable);
-                Assert.False(package10A.IsLatestStableSemVer2);
-
-                Assert.False(package09.IsLatest);
-                Assert.False(package09.IsLatestSemVer2);
-                Assert.False(package09.IsLatestStable);
-                Assert.False(package09.IsLatestStableSemVer2);
-
-                packageRepository.Verify();
-            }
-
-            [Fact]
-            public async Task UpdateIsLatestScenarioForSemVer2PrereleaseAsAbsoluteLatest()
-            {
-                // Arrange
-                var packages = new HashSet<Package>();
-                var packageRegistration = new PackageRegistration { Packages = packages };
-                var semVer2Package = new Package { PackageRegistration = packageRegistration, Version = "1.0.1-alpha.1", IsPrerelease = true, SemVerLevelKey = SemVerLevelKey.SemVer2 };
-                packages.Add(semVer2Package);
-                var package100 = new Package { PackageRegistration = packageRegistration, Version = "1.0.0" };
-                packages.Add(package100);
-                var package10A = new Package { PackageRegistration = packageRegistration, Version = "1.0.0-a", IsPrerelease = true };
-                packages.Add(package10A);
-                var package09 = new Package { PackageRegistration = packageRegistration, Version = "0.9.0" };
-                packages.Add(package09);
-                var packageRepository = new Mock<IEntityRepository<Package>>(MockBehavior.Strict);
-                packageRepository.Setup(r => r.CommitChangesAsync())
-                    .Returns(Task.CompletedTask).Verifiable();
-                var service = CreateService(packageRepository: packageRepository);
-
-                // Act
-                await service.UpdateIsLatestAsync(packageRegistration, true);
-
-                // Assert
-                Assert.True(semVer2Package.IsLatestSemVer2);
-                Assert.False(semVer2Package.IsLatestStableSemVer2);
-                Assert.False(semVer2Package.IsLatest);
-                Assert.False(semVer2Package.IsLatestStable);
-
-                Assert.True(package100.IsLatest);
-                Assert.False(package100.IsLatestSemVer2);
-                Assert.True(package100.IsLatestStable);
-                Assert.True(package100.IsLatestStableSemVer2);
-
-                Assert.False(package10A.IsLatest);
-                Assert.False(package10A.IsLatestSemVer2);
-                Assert.False(package10A.IsLatestStable);
-                Assert.False(package10A.IsLatestStableSemVer2);
-
-                Assert.False(package09.IsLatest);
-                Assert.False(package09.IsLatestSemVer2);
-                Assert.False(package09.IsLatestStable);
-                Assert.False(package09.IsLatestStableSemVer2);
-
-                packageRepository.Verify();
-            }
-
-            [Fact]
-            public async Task UpdateIsLatestScenarioForSemVer2StableAsAbsoluteLatest()
-            {
-                // Arrange
-                var packages = new HashSet<Package>();
-                var packageRegistration = new PackageRegistration { Packages = packages };
-                var semVer2Package = new Package { PackageRegistration = packageRegistration, Version = "1.0.1+metadata", SemVerLevelKey = SemVerLevelKey.SemVer2 };
-                packages.Add(semVer2Package);
-                var package100 = new Package { PackageRegistration = packageRegistration, Version = "1.0.0" };
-                packages.Add(package100);
-                var package10A = new Package { PackageRegistration = packageRegistration, Version = "1.0.0-a", IsPrerelease = true };
-                packages.Add(package10A);
-                var package09 = new Package { PackageRegistration = packageRegistration, Version = "0.9.0" };
-                packages.Add(package09);
-                var packageRepository = new Mock<IEntityRepository<Package>>(MockBehavior.Strict);
-                packageRepository.Setup(r => r.CommitChangesAsync())
-                    .Returns(Task.CompletedTask).Verifiable();
-                var service = CreateService(packageRepository: packageRepository);
-
-                // Act
-                await service.UpdateIsLatestAsync(packageRegistration, true);
-
-                // Assert
-                Assert.True(semVer2Package.IsLatestSemVer2);
-                Assert.True(semVer2Package.IsLatestStableSemVer2);
-                Assert.False(semVer2Package.IsLatest);
-                Assert.False(semVer2Package.IsLatestStable);
-
-                Assert.True(package100.IsLatest);
-                Assert.False(package100.IsLatestSemVer2);
-                Assert.True(package100.IsLatestStable);
-                Assert.False(package100.IsLatestStableSemVer2);
-
-                Assert.False(package10A.IsLatest);
-                Assert.False(package10A.IsLatestSemVer2);
-                Assert.False(package10A.IsLatestStable);
-                Assert.False(package10A.IsLatestStableSemVer2);
-
-                Assert.False(package09.IsLatest);
-                Assert.False(package09.IsLatestSemVer2);
-                Assert.False(package09.IsLatestStable);
-                Assert.False(package09.IsLatestStableSemVer2);
-
-                packageRepository.Verify();
             }
         }
